refactor: extract harbor water/piers carving into harbor_piers.js and wire up import in harbor.js
